feat(arq): wide coherent ARQ window default-ON (16) — 16-bit tone-burst SACK mask (WIRE-BREAKING) + stuck-escape A/B knob
- Tone-burst ACK payload 32->40 bits: frame_mask 8->16 (offsets/CRC-message/
  Hamming blocks re-derived, 27->34 symbols, +26% ACK airtime), uint16_t mask
  chased end-to-end; WIRE-BREAKING: both stations must run this build.
- kToneBurstAckWindowCapFrames 8->16; coherentOFDMWindowOverride() default 16
  for coherent >=R2/3 rungs (the window-16 predicate was a differential-only
  fossil; coherent mods could never reach it). env ULTRA_COHERENT_WINDOW:
  unset->16, =8 legacy, =0 off.
- A/B (50KB faithful gate, sequential, CRC-clean): Good@20 +18/+57/+26%
  (seeds 42/43/7, mean +31%), AWGN@20 3370->3520 NEW RECORD. Default-config
  sanity g42 PASS 1970.
- ULTRA_STUCK_ESCAPE_RETX [2..10] knob (default 5 = unchanged): A/B at 3 is
  seed-dependent (+26% churn seed, -28% good seed) -> not flipped; retx pacing
  is the follow-up lever.
- Docs: CHANGELOG validated entry, KNOWN_BUGS BUG-ACK-STAIRCASE-FADE-BIN was
  STALE (fixed in b85c0e1; occupancy verified 93-100% fast on the gate),
  infrastructure map rows.

+++ src/gui/modem/modem_protocol_binding.hpp
@@ -72,7 +72,7 @@ inline void wireModemToProtocol(ModemEngine& modem,
     // (deliver-once + single GROUP_ACK). THIS is the forwarding ultra_tnc was missing.
     modem.setBurstGroupCallback(
         [&modem, &protocol, hooks](uint16_t group_seq, const std::vector<Bytes>& frames,
-                                   bool all_ok, float quality, uint8_t frame_mask,
+                                   bool all_ok, float quality, uint16_t frame_mask,
                                    bool interleaved, uint8_t group_size) {
             // BUG-ACK-STAIRCASE-FADE-BIN layer 2 (2026-07-01): burst-as-unit delivery
             // bypasses setRawDataCallback, so the frontend's SNR observation STARVED
