+++ src/gui/modem/streaming_decoder.hpp
@@ -183,10 +183,11 @@ using DataSyncAcceptedCallback = std::function<void(float sync_correlation)>;
 // decoded OK. interleaved = the group's bytes were byte-interleaved. When interleaved,
 // only all_ok is meaningful (a partial group is undecodable → whole-burst resend); when
 // NOT interleaved, frame_mask is a true per-frame SACK and the sender resends only the
-// 0-bit frames + refills the burst (the Good/AWGN SR-ARQ path).
+// 0-bit frames + refills the burst (the Good/AWGN SR-ARQ path). 16 bits end-to-end
+// (2026-07-02, matches the tone-burst wire mask kPayloadFrameMaskBits).
 using BurstGroupCallback =
     std::function<void(uint16_t group_seq, const std::vector<Bytes>& frames, bool all_ok,
-                       float quality, uint8_t frame_mask, bool interleaved,
+                       float quality, uint16_t frame_mask, bool interleaved,
                        uint8_t group_size)>;
 
 // StreamingDecoder - Unified RX decoder for all waveform types
@@ -922,7 +923,7 @@ private:
 
     // §15 step 4b: always-on tone-burst ACK monitor. Construction-time tuned
     // for production cadence (longer detect_interval than the unit-test
-    // default to bound CPU; a burst is 675+ ms so a 480 ms cadence still
+    // default to bound CPU; a burst is 850+ ms so a 480 ms cadence still
     // guarantees full coverage). See constructor body.
     ultra::waveform::tone_burst_ack::ToneBurstAckMonitor tone_burst_monitor_;
 };
